requirements.txt
dnspython==2.4.1
motor==3.2.0
pymongo==4.4.1
python-dotenv==1.0.0
black==23.7.0
ipykernel
tensorflow==2.14
polars==0.19.12
Pillow==10.1.0
flask==3.0.0
streamlit
sqlite3